Audit and reduce occurrences of usize.
We want to use usize for counting elements in a buffer or in a
vector.  But for system parameters like "k" or "n" or the number of
counters, we use u32, for consistency across systems.

This patch is incomplete, since we need to change some APIs. I'll do
that in the next commit.